adicionar setup eslint, prettier e jest/react-library
