+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "editor.formatOnSave": false,
+    "editor.codeActionsOnSave": {
+        "source.fixAll.eslint": true 
+    }
+}
