+++ src/components/Form/form.spec.ts
@@ -1,0 +1,4 @@
+it('Primeiro teste', () => {
+  const numero = 10;
+  expect(numero).toBe(10);
+});
