+++ types/propTypes.ts
@@ -1,14 +1,14 @@
 export interface InputFieldProps {
-    label: string;
-    type?: string;
+  label: string;
+  type?: string;
 }
 
 export interface ButtonProps {
-    title: string;
-    type?: 'submit' | 'button' | 'reset';
-    onClick?: React.MouseEventHandler<HTMLButtonElement>;
+  title: string;
+  type?: 'submit' | 'button' | 'reset';
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
 }
 
 export interface FormProps {
-    children: React.ReactNode;
+  children: React.ReactNode;
 }
